Extract out cakery lobby
