+++ HonorAmongThieves/Pages/Index.cshtml
@@ -6,43 +6,7 @@
     <h1 class="display-1"><a href="/Heist">HEIST</a></h1>
     <hr />
     <h1 class="display-1"><a href="/Cakery">CAKERY</a></h1>
-    <h4 class="display-4"><a href="CakeryInstructions.html">Instructions</a></h4>
 </div>
-<hr />
-<h2><u>Existing Lobbies</u></h2>
-@{
-    <h3><u>Heist Lobbies</u></h3>
-    <table border="1">
-        <th>Room ID</th>
-        <th>Updated Time</th>
-        <th>Player Count</th>
-        @foreach (var heistGame in Model.HeistRooms)
-        {
-            <tr>
-                <th>@heistGame.Item1</th>
-                <th>@heistGame.Item2.ToString()</th>
-                <th>@heistGame.Item3.ToString()</th>
-                <th>@heistGame.Item4.ToString()</th>
-            </tr>
-        }
-    </table>
-    <br />
-    <h3><u>Cakery Lobbies</u></h3>
-    <table border="1">
-        <th>Room ID</th>
-        <th>Updated Time</th>
-        <th>Player Count</th>
-        @foreach (var game in Model.CakeryRooms)
-        {
-            <tr>
-                <th>@game.Item1</th>
-                <th>@game.Item2.ToString()</th>
-                <th>@game.Item3.ToString()</th>
-                <th>@game.Item4.ToString()</th>
-            </tr>
-        }
-    </table>
-}
 
 <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/css/bootstrap.min.css" integrity="sha384-GJzZqFGwb1QTTN6wy59ffF1BuGJpLSa9DkKMp0DgiMDm4iYMj70gZWKYbI706tWS" crossorigin="anonymous">
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/js/bootstrap.min.js" integrity="sha384-B0UglyR+jN6CkvvICOB2joaf5I4l3gm9GU6Hc1og6Ls7i6U/mkkaduKaBhlAXv9k" crossorigin="anonymous"></script>
