+++ HonorAmongThieves/Pages/Lobbies/_CakeryLobby.cshtml.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace HonorAmongThieves.Pages.Lobbies
+{
+    public class _CakeryLobbyModel : PageModel
+    {
+        public void OnGet()
+        {
+        }
+    }
+}
